fastapi_stock_analysis/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.1
pydantic==2.9.2
yfinance==0.2.44
pandas==2.2.3
numpy==2.1.3
requests==2.32.3
python-dotenv==1.0.1
anthropic==0.40.0
langchain==0.3.10
langchain-anthropic==0.2.4
